Add base map class
Unlike the Python TCOD tutorial, I'd like the map to be responsible for
as little as possible, and shift that code to the generators themselves.
This way, once the generic classes have been implemented they don't need
to be rewritten in case a generator wants to do something fancy.

The map class will probably need a couple of additional functions to let
outside code interact and change the properties of its tiles.