Fix swapped variables in the Map unit tests
The list comprehension constructor assembles the tile array differently
than I pictured. I thought y was in the upper list, and x was each
nested list, but it's backwards. Which makes sense; first you look up
x, then y. But it means that you can't see the map just by printing out
the array. It's rotated 90 degrees to the left and mirrored from our
starting position of 0,0, which should be in the upper left hand corner.